+++ ChangeLog.md
@@ -6,11 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.7.0] 2018-02-20
+
 ### Added
-- Webworker for paint offloading
+- Webworker for paint offloading (but with postMessage() disabled)
 
 ## Changed
 - Improved update_pixels()
+- Simplified palette
 
 ## [0.6.0] 2018-02-13
 
@@ -71,7 +74,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial creation
 
-[Unreleased]: https://github.com/xyzzy/jsFractalZoom/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/xyzzy/jsFractalZoom/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/xyzzy/jsFractalZoom/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/xyzzy/jsFractalZoom/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/xyzzy/jsFractalZoom/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/xyzzy/jsFractalZoom/compare/v0.3.0...v0.4.0